// src/components/common/paginations/Paginations01.types.ts
import type { Dispatch, SetStateAction } from 'react';

export interface IPaginations01Props {
  dataCount: number;
  startPage: number;
  setStartPage: Dispatch<SetStateAction<number>>;
  activePage: number;
  setActivePage: Dispatch<SetStateAction<number>>;
  eventHook: (pageNumber: any) => void;
}